refactor(server): move the zod json-schema walker into @guren/server
`deriveAgentTools` (RFC 0016 §2) has to derive from the same walker an
OpenAPI document is built from, and it lives in @guren/server. A server
module cannot import @guren/core: core's index is `export * from
'@guren/server'`, so core builds second, and server's declaration build
(`tsc -p tsconfig.build.json`, `paths: {}`, checking on) would look for a
core dist/ that does not exist yet.

So the rule moves down to the package both surfaces can see, and
`@guren/core/internal/zod-compat` / `internal/zod-json-schema` become
re-export shims — every consumer outside @guren/server (@guren/openapi,
@guren/cli) keeps the specifier it already writes. Same shape as
`@guren/server/support/expiry`, re-exported by core's store-utils.

Refs: RFC 0016
