+++ packages/server/package.json
@@ -106,6 +106,14 @@
       "types": "./dist/support/expiry.d.ts",
       "default": "./dist/support/expiry.js"
     },
+    "./internal/zod-compat": {
+      "types": "./dist/internal/zod-compat.d.ts",
+      "default": "./dist/internal/zod-compat.js"
+    },
+    "./internal/zod-json-schema": {
+      "types": "./dist/internal/zod-json-schema.d.ts",
+      "default": "./dist/internal/zod-json-schema.js"
+    },
     "./jsx-runtime": {
       "types": "./dist/jsx-runtime.d.ts",
       "default": "./dist/jsx-runtime.js"
